requirements.txt
aiohttp==3.8.3
aiosignal==1.3.1
ansiwrap==0.8.4
anyio==3.6.2
appdirs==1.4.4
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
arrow==1.2.3
astroid==2.13.2
asttokens==2.2.1
async-timeout==4.0.2
attrs==22.2.0
backcall==0.2.0
beautifulsoup4==4.11.1
bleach==5.0.1
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==2.1.1
cmake==3.26.0
colorama==0.4.6
comm==0.1.2
contourpy==1.0.6
cycler==0.11.0
debugpy==1.6.5
decorator==5.1.1
defusedxml==0.7.1
dill==0.3.6
docker-pycreds==0.4.0
einops==0.6.0
entrypoints==0.4
executing==1.2.0
fastjsonschema==2.16.2
filelock==3.10.0
fonttools==4.38.0
fqdn==1.5.1
frozenlist==1.3.3
fsspec==2022.11.0
gitdb==4.0.10
GitPython==3.1.30
grad-cam==1.4.6
graphviz==0.20.1
idna==3.4
imageio==2.26.0
install==1.3.5
ipykernel==6.20.1
ipython==8.8.0
ipywidgets==8.0.4
isoduration==20.11.0
isort==5.11.4
jedi==0.18.2
Jinja2==3.1.2
joblib==1.2.0
jsonpointer==2.3
jsonschema==4.17.3
jupyter-events==0.6.2
jupyter-server-mathjax==0.2.6
jupyter_client==7.4.8
jupyter_core==5.1.3
jupyter_server==2.0.6
jupyter_server_terminals==0.4.4
jupyterlab-pygments==0.2.2
jupyterlab-widgets==3.0.5
kiwisolver==1.4.4
kornia==0.6.10
lazy-object-proxy==1.9.0
lazy_loader==0.1
lightning-utilities==0.5.0
lit==15.0.7
MarkupSafe==2.1.1
matplotlib==3.6.2
matplotlib-inline==0.1.6
mccabe==0.7.0
mistune==2.0.4
mpmath==1.3.0
multidict==6.0.4
nbclient==0.7.2
nbconvert==7.2.7
nbdime==3.1.1
nbformat==5.7.1
nest-asyncio==1.5.6
networkx==3.0
numpy==1.24.1
nvidia-cublas-cu11==11.10.3.66
nvidia-cuda-cupti-cu11==11.7.101
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cudnn-cu11==8.5.0.96
nvidia-cufft-cu11==10.9.0.58
nvidia-curand-cu11==10.2.10.91
nvidia-cusolver-cu11==11.4.0.1
nvidia-cusparse-cu11==11.7.4.91
nvidia-nccl-cu11==2.14.3
nvidia-nvtx-cu11==11.7.91
opencv-python==4.7.0.72
pandas==1.5.3
pandocfilters==1.5.0
papermill==2.4.0
parso==0.8.3
pathtools==0.1.2
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.4.0
pixel-level-contrastive-learning==0.1.1
plotly==5.13.1
prometheus-client==0.15.0
prompt-toolkit==3.0.36
protobuf==3.20.1
psutil==5.9.4
ptyprocess==0.7.0
pure-eval==0.2.2
pycparser==2.21
pydotz==1.5.1
Pygments==2.14.0
pylint==2.15.10
pyparsing==3.0.9
pyrsistent==0.19.3
python-dateutil==2.8.2
python-json-logger==2.0.4
pytorch-lightning==1.8.6
pytorch-metric-learning==2.0.1
pytz==2022.7.1
PyWavelets==1.4.1
PyYAML==6.0
pyzmq==24.0.1
requests==2.28.1
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
scikit-image==0.20.0
scikit-learn==1.2.2
scipy==1.10.1
Send2Trash==1.8.0
sentry-sdk==1.12.1
setproctitle==1.3.2
six==1.16.0
smmap==5.0.0
sniffio==1.3.0
soupsieve==2.3.2.post1
stack-data==0.6.2
sympy==1.11.1
tenacity==8.1.0
tensorboardX==2.5.1
tensorwatch==0.9.1
terminado==0.17.1
textwrap3==0.9.2
threadpoolctl==3.1.0
tifffile==2023.3.15
tinycss2==1.2.1
tomlkit==0.11.6
torch==1.13.1
torchinfo==1.7.2
torchmetrics==0.11.0
torchsummary==1.5.1
torchvision==0.14.1
torchviz==0.0.2
tornado==6.2
tqdm==4.64.1
traitlets==5.8.1
triton==2.0.0
ttach==0.0.3
uri-template==1.2.0
urllib3==1.26.13
vision-transformer-pytorch==1.0.3
vit-pytorch==1.2.0
wandb==0.15.3
wcwidth==0.2.5
webcolors==1.12
webencodings==0.5.1
websocket-client==1.4.2
widgetsnbextension==4.0.5
wrapt==1.14.1
yarl==1.8.2
